Moved stacktrace
